rename VSMath to VSUtil
